requirements.txt
# FastAPI Server Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Image Processing
Pillow
opencv-python==4.8.1.78
numpy==1.24.3

# HADM Dependencies (from original HADM requirements.txt)
timm==0.5.4
blobfile
mypy
pytest
requests
einops
tensorboardX
deepspeed==0.6.5
scipy
sh
wandb
ftfy
regex
triton
cryptography
ninja

# PyTorch - Will need to be installed separately based on system
# torch>=2.4.0
# torchvision>=0.19.0

# Additional dependencies for detectron2 and HADM
# These will be installed when setting up HADM
# mmcv==1.7.1
# openmim

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==22.3.0
flake8==6.1.0
isort==5.12.0

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# System monitoring
psutil==5.9.6 